+++ src/ChaosTools.jl
@@ -17,15 +17,14 @@ include("orbitdiagrams/discrete_diagram.jl")
 include("orbitdiagrams/poincare.jl")
 include("orbitdiagrams/produce_orbitdiagram.jl")
 
-include("basins/attractor_mapping.jl")
-include("basins/attractor_mapping_proximity.jl")
-include("basins/attractor_mapping_featurizing.jl")
-include("basins/attractor_mapping_recurrences.jl")
+include("basins/dict_utils.jl")
+include("basins/mapping/attractor_mapping.jl")
 include("basins/basins_of_attraction.jl")
 include("basins/basins_utilities.jl")
 include("basins/fractality_of_basins.jl")
 include("basins/tipping.jl")
 include("basins/sampler.jl")
+include("basins/continuation/basins_fractions_continuation_api.jl")
 
 include("dimensions/linear_regions.jl")
 include("dimensions/generalized_dim.jl")
@@ -55,6 +54,8 @@ include("chaosdetection/partially_predictable.jl")
 include("chaosdetection/testchaos01.jl")
 
 # Exports
+# TODO: It is probably better to put all exports here rather than in
+# each individual file. Or is it...? Not sure yet.
 export lyapunovspectrum, lyapunov, local_growth_rates, lyapuov_from_data