+++ src/basins/mapping/attractor_mapping.jl
@@ -1,3 +1,6 @@
+# Definition of the attracting mapping API and exporting
+# At the end it also includes all files related to mapping
+
 export AttractorMapper,
     AttractorsViaRecurrences,
     AttractorsViaRecurrencesSparse,
@@ -63,7 +66,7 @@ Initial conditions to use are defined by `ics`. It can be:
   See [`statespace_sampler`](@ref) to generate such functions.
 
 The returned arguments are `fs`.
-If `ics` is a `Dataset` then the `labels` of each initial and roughly approximated
+If `ics` is a `Dataset` then the `labels` of each initial condition and roughly approximated
 attractors are also returned: `fs, labels, attractors`.
 
 The output `fs` is a dictionary whose keys are the labels given to each attractor
@@ -79,14 +82,14 @@ See [`AttractorMapper`](@ref) for all possible `mapper` types.
 * `show_progress = true`: Display a progress bar of the process.
 """
 function basins_fractions(mapper::AttractorMapper, ics::Union{AbstractDataset, Function};
-        show_progress = true, N = 1000
+        show_progress = true, N = 1000, additional_fs::Dict = Dict(),
     )
     used_dataset = ics isa AbstractDataset
     N = used_dataset ? size(ics, 1) : N
     if show_progress
         progress=ProgressMeter.Progress(N; desc="Mapping initial conditions to attractors:")
     end
-    fs = Dict{Int, Float64}()
+    fs = Dict{Int, Int}()
     used_dataset && (labels = Vector{Int}(undef, N))
     # TODO: If we want to parallelize this, then we need to initialize as many
     # mappers as threads. Use a `threading` keyword and `deepcopy(mapper)`
@@ -97,6 +100,10 @@ function basins_fractions(mapper::AttractorMapper, ics::Union{AbstractDataset, F
         used_dataset && (labels[i] = label)
         show_progress && next!(progress)
     end
+    # the non-public-API `additional_fs` is used in the continuation methods
+    additive_dict_merge!(fs, additional_fs)
+    N = N + (isempty(additional_fs) ? 0 : sum(values(additional_fs)))
+    # Transform count into fraction
     ffs = Dict(k => v/N for (k, v) in fs)
     if used_dataset
         attractors = extract_attractors(mapper, labels, ics)
@@ -135,9 +142,9 @@ the partitioning happens directly on the hyperplane the Poincaré map operates o
 corresponding to the state space partitioning indicated by `grid`.
 """
 function basins_of_attraction(mapper::AttractorMapper, grid::Tuple; kwargs...)
-    basins = zeros(Int16, map(length, grid))
+    basins = zeros(Int32, map(length, grid))
     I = CartesianIndices(basins)
-    A = Dataset([generate_ic_on_grid(grid, I[i]) for i in 1:length(I)])
+    A = Dataset([generate_ic_on_grid(grid, i) for i in I])
     fs, labels, attractors = basins_fractions(mapper, A; kwargs...)
     vec(basins) .= vec(labels)
     return basins, attractors
@@ -151,3 +158,10 @@ end
         @inbounds return SVector{$B, Float64}($(gens...))
     end
 end
+
+#########################################################################################
+# Includes
+#########################################################################################
+include("attractor_mapping_proximity.jl")
+include("attractor_mapping_recurrences.jl")
+include("attractor_mapping_featurizing.jl")