+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # 2.10
+- New function `basins_fractions_continuation` that calculates basins fractions and how these change versus a parameter (given a continuation method)
+- New basins fraction continuation method `RecurrencesSeededContinuation` that utilizes a brand new algorithm to continuate basins fractions of arbitrary systems.
+- `match_attractor_ids!` has been fully overhauled to be more flexible, allow more ways to match, and also allow arbitrary user-defined ways to match.
+- New function `match_basins_ids!` for matching the output of `basins_of_attraction`.
+- New attractor mapping algorithm `AttractorsViaRecurrencesSparse` that uses sparse arrays to find attractors of arbitrarily high dimensional dynamical systems, eliminating the main drawback of `AttractorsViaRecurrences`.
 - Clustering (used in `AttractorsViaFeaturizing`) has been completely overhauled. Now, a `ClusteringConfig` instances must be created and then passed on to `AttractorsViaFeaturizing`.
 - `AttractorsViaFeaturizing` no longer has keywords about clustering.
 - A new function `cluster_features` is exposed to the user.